requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.3.2 